dernier commit groupomania

+++ client/src/components/Profil/UpdateDeleteProfil.js
@@ -21,48 +21,49 @@ const UpdateProfil = (props) => {
     }
 
     const handleUpdate = () => {
-       dispatch(updateBio(userData.id, bio));
-       setUpdateForm(false);
+        dispatch(updateBio(userData.id, bio));
+        setUpdateForm(false);
     }
 
     return (
         <div className='profil-container'>
             <LeftNav />
-            <h1> Profil de {userData.firstname + " "+userData.name}</h1>
+            <h1> Profil de {userData.firstname + " " + userData.name}</h1>
             <div className='update-container'>
-             <div className='left-part'>
-             <h3>Photo de profil</h3>
-             <img src={userData.picture} alt="user-pic" />  
-             <UploadImg />
-             </div>
-             <div className='right-part'>
-                <div className='bio-update'>
-                   <h3>Biographie</h3>
-                   {updateForm === false && (
-                       <>
-                       <p onClick={() => setUpdateForm(!updateForm)}>{userData.bio}</p>
-                       <button onClick={() => setUpdateForm(!updateForm)}>Modifier la biographie</button>
-                       </>
-                   )}
-                   {updateForm && (
-                      <>
-                       <textarea type="text" 
-                       defaultValue={userData.bio} 
-                       onChange={(e) => setBio(e.target.value)}>    
-                       </textarea>
-                       <button onClick={handleUpdate}>Valider les modifications</button>  
-                      </> 
-                   )} 
+                <div className='left-part'>
+                    <h2>Photo de profil</h2>
+                    <br />
+                    <img src={userData.picture} alt="user-pic" />
+                    <UploadImg />
                 </div>
-                <h4>Membre depuis le : {dateParser(userData.createdAt)}</h4>
-                <button onClick={ () => {
-                    if (window.confirm('êtes vous sûr de vouloir supprimer votre compte Groupomania ?'))
-                    { deleteProfilUser() }
-                }}>Supprimer le compte
+                <div className='right-part'>
+                    <div className='bio-update'>
+                        <h3>Biographie</h3>
+                        {updateForm === false && (
+                            <>
+                                <p onClick={() => setUpdateForm(!updateForm)}>{userData.bio}</p>
+                                <br />
+                                <button onClick={() => setUpdateForm(!updateForm)}>Modifier la biographie</button>
+                            </>
+                        )}
+                        {updateForm && (
+                            <>
+                                <textarea type="text"
+                                    defaultValue={userData.bio}
+                                    onChange={(e) => setBio(e.target.value)}>
+                                </textarea>
+                                <button onClick={handleUpdate}>Valider les modifications</button>
+                            </>
+                        )}
+                    </div>
+                    <h4>Membre depuis le : {dateParser(userData.createdAt)}</h4>
+                    <button onClick={() => {
+                        if (window.confirm('êtes vous sûr de vouloir supprimer votre compte Groupomania ?')) { deleteProfilUser() }
+                    }}>Supprimer le compte
 
-                </button>
-             </div>
-             
+                    </button>
+                </div>
+
             </div>
         </div>
     );
